on auth change, rerender router
