+++ src/routes/User.jsx
@@ -1,23 +1,9 @@
-import { useEffect } from 'react';
-import { onAuthStateChanged } from 'firebase/auth';
-import { auth } from '../utils/firebase';
 import useCurrentUserData from '../utils/useCurrentUserData';
 import SignIn from '../pages/SignIn';
 import Profile from '../pages/Profile';
 
 const User = () => {
-  const { userData, setUserData } = useCurrentUserData();
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      if (user) {
-        setUserData({ username: user.displayName, email: user.email, userid: user.uid });
-      } else {
-        setUserData(undefined);
-      }
-    });
-    return () => unsubscribe();
-  }, []);
+  const { userData } = useCurrentUserData();
 
   if (userData) {
     return <Profile />;
